Add: Initial prototype
Adds prototype of a task page. It can be shared and synced via export as
html which  allows offline usage.

It is utilizing a localstorage for things that have not been exported
yet.

The idea is to also create a backend to allow usage without JS but
online and a sync functionality for the events. This would allow usage
when there is a backend available without having to export and reload a
page.

There is a lot missing yet:
- deletion of tasks
- labels with bg colors for tasks
- the possibility to change a board
- the possibility to add multiple boards
- the possibility to manage labels for boards
- styling
- a backend for convenience
- attachments for taks
- allow bold, italic, strikethrough for description